과연?!

+++ src/main/java/com/team/berp/domain/CompanyOrder.java
@@ -56,7 +56,7 @@ public class CompanyOrder {
     private String note;
 
     @Column
-    private Long orderQty;
+    private Integer orderQty;
     
     @Column(name = "order_num")
     private String orderNum;
